refactor(MenuValidator): 도메인 서비스 인터페이스로 변경
